Moved packet parsing and generation into utils.rs for reuse
